Chore : 카테고리 코드 enum 타입 분리
